Added settings & support for light/dark themes. The Settings page can be accessed from a temporary button in the Main Menu (proper settings button to be added later).
The Settings activity is a black-box representation of the root_preferences.xml file, where the different preferences are defined.

Themes can be toggled from a Switch button in the SettingsActivity and are saved to Shared Preferences. I made the theme more complete as possible so that, from now on, all elements should default to the theme-dependent colour scheme without the need for additional configuration (e.g. buttons will have a background colour of primaryColor and a text colour equal to the window background color.
Colours for Dark Mode can be accessed in the new colours-night.xml file.
Removed the white background of the ZSL logo so it would work with the 2 themes